Migrate styles from scss to vanilla-extract css in js
The base and reset styles have been migrated from scss to a CSS-in-JS solution (vanilla-extract). The scss files have been deleted, converted, and their structure has been split in two new files: base.css.ts and vars.css.ts. This change improves the maintainability and scalability of our styling system. Reusable styles are now handled in a CSS-in-JS file, ensuring variable values to be type-safe and easier to manage.