Add RootLayout component and style for application layout
This commit introduces a RootLayout component that serves as the base layout for the application. It incorporates a SideNavigation, some metadata and tracks Google Analytics. The layout styling was also added in a separate CSS module. The aim of these changes is to structure the application and establish a clean and easy-to-navigate interface.
Constants file was also moved to app/constants for better organization.