Move main style to layout.css and wrap children in main tag
Moved the 'main' style from the 'page.css.ts' file to the 'layout.css.ts' file. Also, wrapped the children in a 'main' HTML tag enclosed in the 'layout.tsx' file. This change ensures a more organized structure of the application, and makes the 'main' style reusable across different components.