Add Contentful CLI scripts and dependencies
This commit adds two new scripts related to Contentful: cf-export and cf-generate-types, and adds related dependencies (contentful-cli, cf-content-types-generator, dotenv) to package.json and package-lock.json. The cf-export script is for exporting content from Contentful, and the cf-generate-types script is for generating TypeScript typedefs for the Contentful schema. They are essential for maintaining content and types consistency between Contentful and the codebase.