Updated types in types/TagType.ts and types/EntryType.ts
Simplified the check for tag and medium in a slug by directly accessing slug instead of fields.slug - makes the code cleaner and easier to read.
In types/TagType.ts, added an id field to the TagType to uniquely identify each tag.
In types/EntryType.ts, made changes to clarify the type definitions, removed the fields metaInfo and slid, and modified the type definitions of medium and keyvisual for a more accurate representation of the data types. These changes in types make the type's usability more explicit and easier to use across the application.