Recreate LinkCard component for optimized usability
The LinkCard component was completely refactored to improve its user interface and functionality. The styles (SCSS) and Typescript files were deleted and replaced with new CSS.ts and TSX files as part of implementing the vanilla-extract approach, which aligns better with the existing codebase and style, and streamlines the development process.

The new LinkCard now uses Next.js's Image component instead of an img tag for better image optimization. The component's layout has also been improved using CSS grid, which allows for more flexible and efficient layout structure. This change simplifies the rendering process, enhances the reusability of the component and improves the user experience with faster load times.