Add remote image sources to next.config.js
In this update, `images` configuration has been added in the `next.config.js` file to allow the use of remote image sources. This change is crucial to allow the application to fetch images from various external servers' defined protocols, hostnames, and ports, thereby improving the flexibility and extendability of content fetching scenarios in the application.