Refactor MenuButton component and move to common
Deleted the existing MenuButton component and its stylings from components/header and re-implemented it inside app/components/common. The reason for this refactor is to make this component reusable across other parts of the application. It's also been migrated to use vanilla-extract for CSS in JS styling for better component scope styling, performance, and type safety.