Update styles for container query
This commit updates several styles on multiple pages to improve layout control. This includes changing of aspect ratio of EntryArticle as well as adding responsive breakpoints for different viewports. Navbar has been updated to improve legibility and hierarchy of information. The root container in "layout.css.ts" was added to center the wrapper and other elements horizontally in the viewport. The "vars.css.ts" file was updated to add a new spacing of 80px.