Remove store.ts file
This commit removes the 'store.ts' file from the codebase.
This file was managing the state of the navigation in the application and was using Redux.
However, due to a structural overhaul of how we handle component state throughout the app, this file is no longer required. This simplifies the state management and reduces the overall complexity of the system.