Add auto-close functionality to overlay menu
The overlay menu now has an extra `onChangePage` property that will automatically close the menu when navigating away from the page. This added functionality improves user experience by reducing the need for additional clicks and simplifies navigation. It was implemented by calling `closeModal` inside `onChangePage` when a link is clicked.