Move generated types folder path to app level
The generated types from contentful were previously at the root level. It was decided to move them inside the 'app' directory for better project structure organization and to make it more intuitive for other developers when navigating the project. All corresponding import paths and script in package.json were updated.