Refactor LinkCard component to handle null metaInfo
The LinkCard component now handles null metaInfo fields in a streamlined manner. Removed initial condition which was checking for null metaInfo. Instead, opted for optional chaining within the component render. This gives the ability to display a LinkCard even without ogImage and ogTitle present, as the linkUrl will always be