Simplify Google Analytics component
Removed unnecessary usePathname, useSearchParams from GoogleAnalytics.tsx. The earlier approach was tracking pathname and searchParams to create url for each page view, which made the code more complex and harder to maintain. It turns out that this is not necessary as the Google Analytics script handles this itself. This change simplifies the code and makes it