Add "About" link to OverlayNavigation component
The OverlayNavigation component has been updated to include a new "About" link. This new link fits into the current navigation structure and redirects to the "/about" page. This change has been done to allow users to easily navigate to the "About" page from anywhere in the application. Additionally, the paddingBlock value in the OverlayNavigation CSS module has been increased for better visual arrangement.