Refactor metadata fetch in various page components
The metadata fetch method (`getMetaDataForEntryDataList`) which previously was invoked during the fetching of entries has been moved to the component level. The reason for making this change was to make components responsible for their data fetching in order to improve usability, maintainability, and testability. This change affects entry, medium, tag, and main page components.