Update navigation bar style for better UX
The Navigation.css.ts file was updated to improve the user navigation experience. The main navigation bar is now sticky at the top of the viewport with a fixed height of 80px in a full width. These styling changes allow users to have continuous access to the navigation bar, enhancing user experience. In screens smaller than 800px, navigation bar resumes relative position and auto width/height for a responsive and user-friendly site in various screen sizes.