Update package dependencies and remove development flags
This commit updates the project's package dependencies including adding the "next-pwa" package and removes the "dev" flag in the settings for packages including "@babel/code-frame", "@babel/core", "@babel/compat-data", and others. These changes are generally important to ensure that the project is using the most up-to-date and secure versions of its dependencies. The removal of development flags from some packages could suggest changes in the project's compilation and distribution process, indicating a translation from a development project to a production-ready state.