Added PWA support and enabled React strict mode
The code changes first involve including 'next-pwa' package which enables the application to function as a Progressive Web App (PWA). The destination folder for the generated service worker file is set as the 'public' folder.

Additionally, React strict mode is enabled for the entire application to help highlight potential problems in its components. This forms part of the general optimisation strategy, aiming to improve the development and the eventual user experience.