Change EntryArticle component from article to div & add hover effects
The EntryArticle component was adjusted from an article element to a div in EntryArticle.tsx. This is to ensure better support across browsers and consistency with other component structuring. Additionally, selectors were added in EntryArticle.css.ts that change elements' color to primary when the linkEntry is hovered. This enhances user interaction by visually indicating the element under the pointer.