Make listTitle a required prop in EntryList component
Previously, the listTitle prop in the EntryList component was optional. In cases where a listTitle was not provided, no title was displayed for the list. However, having a title improves user experience by giving context to the list content. So, this change makes listTitle a required prop for the EntryList component. Adjustments have been made in the relevant pages where the EntryList component is used, to pass in a listTitle.