Refactor published date layout in EntryList
The layout of the published date in EntryList component was adjusted. Previously, we were using <time> element directly but it's wrapped inside a <p> element now. The change allows the semantics of the time element to be more evident and render more consistently across different browsers. The formatting adjustments were also made on the corresponding CSS files, removing unnecessary "display: block" property, and adding an intentional margin on top.