Add Storybook story for EntryList component
Added a new Storybook story for the EntryList component to improve development experience by providing a visual representation to test and debug the component. This story simulates a list of entries with predefined data that can also be used as a reference for how data should be shaped for the component.