Update LinkCard and EntryList stories
The storybook for LinkCard and EntryList components has been updated. Replaced the previously used placeholder image and title with a more descriptive sample content in the LinkCard's metaInfo. Also, the titles of sections under LinkCard have been changed to convey the situation when parent element width is less than or more than 500px. This helps in better understanding the component behavior in constrained environments.